requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
aio-pika==9.3.0
motor==3.1.1
pytest==7.4.4
pytest-asyncio==0.21.1 